% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_white.R
\docType{data}
\name{whiteshark}
\alias{whiteshark}
\title{White Shark Occurrences}
\format{
data.frame with thre columns
\describe{
\item{x}{longitude (decimal)}
\item{y}{latitude (decimal)}
}
}
\source{
OBIS
}
\usage{
whiteshark
}
\description{
Global georeferenced occurrences for the great white shark. Thinned to avoid duplicated coordinates
}
\examples{
data(data_white)

}
\keyword{datasets}
